fix: correct data type of averageComparisons and initialize static variables in SearchingAlgo class

+++ TD2/Exo1/searching_algo.h
@@ -12,7 +12,7 @@ class SearchingAlgo {
         int numberComparisons;
         static int totalComparisons;
         static int totalSearch;
-        static int averageComparisons;
+        static double averageComparisons;
 
         virtual int search(const std::vector<int> elements, int searchKey) = 0;
